+++ src/main/java/example/cashcard/CashCardRepository.java
@@ -5,10 +5,10 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-interface CashCardRepository extends CrudRepository<CashCard, Long>, PagingAndSortingRepository<CashCard, Long> {
-    CashCard findByIdAndOwner(Long id, String owner);
+interface CashCardRepository extends CrudRepository<CashCard, Integer>, PagingAndSortingRepository<CashCard, Integer> {
+    CashCard findByIdAndOwner(int id, String owner);
 
-    boolean existsByIdAndOwner(Long id, String owner);
+    boolean existsByIdAndOwner(int id, String owner);
 
     Page<CashCard> findByOwner(String owner, PageRequest pageRequest);
 }
